new file:   Data/conservation ,cell,ri=5,ro=10,dt=0.0001,t=0.5,v=0.5,t_tot=10,interval=0.1.avi 	new file:   Data/conservation ,cell,ri=5,ro=10,dt=0.0001,t=0.5,v=0.5,t_tot=5,interval=0.1.avi 	new file:   Data/conservation ,cell,ri=5,ro=15,dt=0.0001,t=0.4,v=0.5,t_tot=5,interval=0.1.avi 	new file:   Data/conservation ,cell,ri=5,ro=15,dt=0.0001,t=1,v=0.5,t_tot=5,interval=0.1.avi 	new file:   Data/conservation ,ri=5,ro=15,dt=0.0001,t=1,v=0.5,t_tot=5,interval=0.1.avi 	new file:   Data/conservation, adaptive ,cell,ri=5,ro=10,dt=0.0001,t=0.5,v=0.5,t_tot=5,interval=0.1.avi 	deleted:    Data/data3.asv 	modified:   Data/data3.m 	new file:   Data/energy_conservation/conservation ,cell,ri=5,ro=10,dt=0.0001,t=0.5,v=0.5,t_tot=10,interval=0.1.csv 	new file:   Data/energy_conservation/conservation ,cell,ri=5,ro=10,dt=0.0001,t=0.5,v=0.5,t_tot=5,interval=0.1.csv 	new file:   Data/energy_conservation/conservation ,cell,ri=5,ro=10,dt=0.0001,t=1,v=0.5,t_tot=5,interval=0.1.csv 	deleted:    Data/energy_conservation/conservation ,cell,ri=5,ro=15,dt=0.0001,t=0.2,v=0.2,t_tot=5,interval=0.1.csv 	deleted:    Data/energy_conservation/conservation ,cell,ri=5,ro=15,dt=0.0001,t=0.4,v=1,t_tot=5,interval=0.1.csv 	deleted:    Data/energy_conservation/conservation ,cell,ri=5,ro=15,v=1,t=0.4,t_tot=5,interval=0.1,dt=0.00004.csv 	deleted:    Data/energy_conservation/conservation ,cell,ri=5,ro=15,v=1,t=0.4,t_tot=5,interval=0.1,dt=0.0001.csv 	deleted:    Data/energy_conservation/conservation ,cell,ri=5,ro=20,v=1,t=0.4,t_tot=5,interval=0.1,dt=0.0001.csv 	deleted:    Data/energy_conservation/conservation ,cell,ri=5,ro=20,v=1,t=0.5,t_tot=5,dt=0.0001.csv 	deleted:    Data/energy_conservation/conservation ,cell,static.csv 	new file:   Data/energy_conservation/conservation, adaptive ,cell,ri=5,ro=10,dt=0.0001,t=0.5,v=0.5,t_tot=5,interval=0.1.csv 	modified:   Data/plotEnergyComponents.m 	modified:   Data/specific_time.m 	modified:   Data/updateParticleVisualization.m 	deleted:    Data/vibration_energy.asv 	modified:   Data/vibration_energy.m 	modified:   bin/energy_conservation.exe 	modified:   bin/energy_conservation_cell_2.exe 	modified:   energy_conservation.cpp 	modified:   energy_conservation_cell_2.cpp 	new file:   include/Adaptive_Verlet.h 	modified:   include/Constrain.h 	modified:   include/Energy.h 	modified:   include/LJForce.h 	modified:   include/Particle.h 	modified:   include/Save.h 	modified:   include/Simulate_Triangular.h 	modified:   include/Simulate_Triangular_Cell.h 	modified:   include/Triangular_Distribution.h 	modified:   include/Vec2.h

+++ include/Vec2.h
@@ -6,7 +6,7 @@
 #include <cmath>
 
 struct Vec2 {
-    float x, y;
+   double x, y;
 
     Vec2 operator+=(const Vec2& other) {
         x += other.x;
@@ -14,7 +14,7 @@ struct Vec2 {
         return *this;
     }
 
-    Vec2 operator*(float scalar) const {
+    Vec2 operator*(double scalar) const {
         return {x * scalar, y * scalar};
     }
 
@@ -23,17 +23,17 @@ struct Vec2 {
     }
     
 
-    float length() const {
+   double length() const {
         return std::sqrt(x * x + y * y);
     }
 
     Vec2 normalized() const {
-        float len = length();
+        double len = length();
         return (len > 0) ? Vec2{x / len, y / len} : Vec2{0, 0};
     }
 
     // 构造函数
-    Vec2(float x = 0, float y = 0) : x(x), y(y) {}
+    Vec2(double x = 0, double y = 0) : x(x), y(y) {}
 
     // 重载负号运算符
     Vec2 operator-() const {
